사다리게임 객체구조 개선을 위한 파일이동
