--- src/main/java/net/chandol/domain/bridge/BridgesSupplier.java
@@ -1,7 +1,0 @@
-package net.chandol.domain.bridge;
-
-import net.chandol.value.Positive;
-
-public interface BridgesSupplier {
-    Bridges createBridges(Positive lineSize, Positive height);
-}
