+++ src/main/java/net/chandol/ladder/v1/domain/bridge/BridgesSupplier.java
@@ -1,0 +1,7 @@
+package net.chandol.ladder.v1.domain.bridge;
+
+import net.chandol.ladder.v1.value.Positive;
+
+public interface BridgesSupplier {
+    Bridges createBridges(Positive lineSize, Positive height);
+}
